refactor: Move role evaluation to common
It's common functionality anyways
